add slanted div text color calculations
I have added logic to calculate and determine whether text falls within the light or dark area. This allows us to display both light and dark text over the two divs. My calculations are not perfect and it is causing issues with visibility of some characters. This styling is also flawed due to difficulty reading cutoff characters.

I also added a change in the tsconfig to allow for iterating over html elements with modern syntax.