remove buggy split display
While I had fun making it, I could never get the split display to work as intended. Even the best-case-scenario version would not be very accessible and would make it difficult to read. I am removing it now, but leaving it in the codebase just in case we wish to use it someday.