scale up svg and fix formatting to support new scale
On some browsers/devices the SVG in the banner was ignoring the w-screen directive and not filling the screen. I have scaled up the SVG to fix this issue. I also scaled up the entire SVG to look a bit better on all devices. I increased the negative y margin to support the larger size. This may need testing as this type of hard-coding tends to be fragile. Testing on 1080 monitor and multiple responsive settings. Should be tested to support 1440 as well, may need new media queries to support this screen size.